Refactored the menu bar component to move all modal contents to their own respetive component
