+++ src/Viewer/components/MenuBar/MenuBar.js
@@ -1,15 +1,17 @@
 import React, {useContext, useRef, useState} from "react";
 
 import PropTypes from "prop-types";
-import {Button, Form, Modal, ProgressBar, Table} from "react-bootstrap";
+import {Modal, ProgressBar} from "react-bootstrap";
 import {ChevronDoubleLeft, ChevronDoubleRight, ChevronLeft, ChevronRight,
-    FileText, Folder, Gear, Keyboard, Moon, Sun} from "react-bootstrap-icons";
+    FileText, Folder, Gear, Keyboard} from "react-bootstrap-icons";
 
 import {THEME_STATES} from "../../../ThemeContext/THEME_STATES";
 import {ThemeContext} from "../../../ThemeContext/ThemeContext";
 import MODIFY_PAGE_ACTION from "../../services/MODIFY_PAGE_ACTION";
 import STATE_CHANGE_TYPE from "../../services/STATE_CHANGE_TYPE";
 import {EditableInput} from "./EditableInput/EditableInput";
+import {HelpModal} from "./HelpModal/HelpModal";
+import {SettingsModal} from "./SettingsModal/SettingsModal";
 
 import "./MenuBar.scss";
 
@@ -45,23 +47,20 @@ MenuBar.propTypes = {
  * @param {LoadFileCallback} loadFileCallback
  * @return {JSX.Element}
  */
-export function MenuBar ({
-    logFileState, fileMetaData, loadingLogs, changeStateCallback, loadFileCallback,
-}) {
-    const {theme, switchTheme} = useContext(ThemeContext);
+export function MenuBar ({logFileState, fileMetaData, loadingLogs,
+    changeStateCallback, loadFileCallback}) {
+    const {theme} = useContext(ThemeContext);
 
-    const [eventsPerPage, setEventsPerPage] = useState(logFileState.pages);
+    // Modal State Variables
     const [showSettings, setShowSettings] = useState(false);
-    const [showHelp, setShowHelp] = useState(false);
-
     const handleCloseSettings = () => setShowSettings(false);
     const handleShowSettings = () => setShowSettings(true);
 
+    const [showHelp, setShowHelp] = useState(false);
     const handleCloseHelp = () => setShowHelp(false);
     const handleShowHelp = () => setShowHelp(true);
 
-    const inputFile = useRef(null);
-
+    // Page navigation functions
     const goToFirstPage = () => {
         if (logFileState.page !== 1) {
             changeStateCallback(STATE_CHANGE_TYPE.page, {action: MODIFY_PAGE_ACTION.firstPage});
@@ -89,52 +88,6 @@ export function MenuBar ({
         });
     };
 
-    // File functions
-    const openFile = () => {
-        inputFile.current.click();
-    };
-
-    const loadFile = (e) => {
-        loadFileCallback(e.target.files[0]);
-    };
-
-    // Modal Functions
-    const getModalClass = () => {
-        return (THEME_STATES.LIGHT === theme)?"modal-light":"modal-dark";
-    };
-
-    const saveModalChanges = (e) => {
-        // TODO Can't backspace 0 from the number input
-        // TODO What is the maximum number of events monaco can support?
-        e.preventDefault();
-        handleCloseSettings();
-        changeStateCallback(STATE_CHANGE_TYPE.pageSize, {pageSize: eventsPerPage});
-        localStorage.setItem("pageSize", String(eventsPerPage));
-    };
-
-    const closeModal = () => {
-        handleCloseSettings();
-    };
-
-    const openModal = () => {
-        handleShowSettings();
-        setEventsPerPage(logFileState.pageSize);
-    };
-
-    const getThemeIcon = () => {
-        if (THEME_STATES.LIGHT === theme) {
-            return (
-                <Moon className="cursor-pointer" title="Set Light Mode"
-                    onClick={() => switchTheme(THEME_STATES.DARK)}/>
-            );
-        } else if (THEME_STATES.DARK === theme) {
-            return (
-                <Sun className="cursor-pointer" title="Set Dark Mode"
-                    onClick={() => switchTheme(THEME_STATES.LIGHT)}/>
-            );
-        }
-    };
-
     const getPageNav = () => {
         return (
             <>
@@ -160,6 +113,23 @@ export function MenuBar ({
         );
     };
 
+    // File State and functions
+    const inputFile = useRef(null);
+    const openFile = () => {
+        inputFile.current.click();
+    };
+    const loadFile = (e) => {
+        loadFileCallback(e.target.files[0]);
+    };
+
+    // Callback to save the changes in the settings modal
+    const saveSettingsChanges = (eventsPerPage) => {
+        // TODO What is the maximum number of events monaco can support?
+        changeStateCallback(STATE_CHANGE_TYPE.pageSize, {pageSize: eventsPerPage});
+        localStorage.setItem("pageSize", String(eventsPerPage));
+    };
+
+    // Loading bar functions and variables
     const loadingBarHeight = "3px";
     const getLoadingBar = () => {
         return (loadingLogs)
@@ -168,7 +138,6 @@ export function MenuBar ({
     };
 
     // TODO make file icon a button to open modal with file info
-    // TODO Move modals into their own component
     return (
         <>
             <div className="viewer-header" data-theme={theme}>
@@ -183,7 +152,7 @@ export function MenuBar ({
                     <div className="menu-right">
                         {getPageNav()}
                         <div className="menu-divider"></div>
-                        <div className="menu-item menu-item-btn" onClick={openModal}>
+                        <div className="menu-item menu-item-btn" onClick={handleShowSettings}>
                             <Gear/>
                         </div>
                         <div className="menu-divider"></div>
@@ -204,123 +173,17 @@ export function MenuBar ({
             </div>
 
             <Modal show={showSettings} className="border-0" onHide={handleCloseSettings}
-                contentClassName={getModalClass()}>
-                <Modal.Header className="modal-background border-0" >
-                    <div className="float-left">
-                        App Settings
-                    </div>
-                    <div className="float-right">
-                        {getThemeIcon()}
-                    </div>
-                </Modal.Header>
-                <Modal.Body className="modal-background p-3 pt-1" >
-                    <label className="mb-2">Log Events per Page</label>
-                    <Form onSubmit={saveModalChanges}>
-                        <Form.Control type="number"
-                            value={eventsPerPage}
-                            onChange={(e) => setEventsPerPage(Number(e.target.value))}
-                            className="input-sm num-event-input" />
-                    </Form>
-                </Modal.Body>
-                <Modal.Footer className="modal-background border-0" >
-                    <Button className="btn-sm" variant="success" onClick={saveModalChanges}>
-                        Save Changes
-                    </Button>
-                    <Button className="btn-sm" variant="secondary" onClick={closeModal}>
-                        Close
-                    </Button>
-                </Modal.Footer>
+                contentClassName={(THEME_STATES.LIGHT === theme)?"modal-light":"modal-dark"}>
+                <SettingsModal
+                    handleClose={handleCloseSettings}
+                    saveChangesCallback={saveSettingsChanges}
+                    initialEventsPerPage={logFileState.pageSize}
+                ></SettingsModal>
             </Modal>
 
-
             <Modal show={showHelp} className="help-modal border-0" onHide={handleCloseHelp}
-                contentClassName={getModalClass()} data-theme={theme}>
-                <Modal.Header className="modal-background" >
-                    <div className="float-left">
-                        Keyboard Shortcuts
-                    </div>
-                </Modal.Header>
-                <Modal.Body className="modal-background p-3 pt-2" >
-                    <Table borderless style={{fontSize: "15px"}} >
-                        <thead>
-                            <tr>
-                                <th>Action</th>
-                                <th>Windows</th>
-                                <th>macOS</th>
-                            </tr>
-                        </thead>
-                        <tbody>
-                            <tr>
-                                <td>Focus on Editor</td>
-                                <td>
-                                    <kbd>`</kbd>
-                                </td>
-                                <td>
-                                    <kbd>`</kbd>
-                                </td>
-                            </tr>
-                            <tr>
-                                <td>Next Page</td>
-                                <td>
-                                    <kbd>CTRL</kbd>+<kbd>]</kbd>
-                                </td>
-                                <td>
-                                    <kbd>⌘</kbd>+<kbd>]</kbd>
-                                </td>
-                            </tr>
-                            <tr>
-                                <td>Prev Page</td>
-                                <td>
-                                    <kbd>CTRL</kbd>+<kbd>[</kbd>
-                                </td>
-                                <td>
-                                    <kbd>⌘</kbd>+<kbd>[</kbd>
-                                </td>
-                            </tr>
-                            <tr>
-                                <td>First Page</td>
-                                <td>
-                                    <kbd>CTRL</kbd>+<kbd>,</kbd>
-                                </td>
-                                <td>
-                                    <kbd>⌘</kbd>+<kbd>,</kbd>
-                                </td>
-                            </tr>
-                            <tr>
-                                <td>Last Page</td>
-                                <td>
-                                    <kbd>CTRL</kbd>+<kbd>.</kbd>
-                                </td>
-                                <td>
-                                    <kbd>⌘</kbd>+<kbd>.</kbd>
-                                </td>
-                            </tr>
-                            <tr>
-                                <td>Top of Page</td>
-                                <td>
-                                    <kbd>CTRL</kbd>+<kbd>U</kbd>
-                                </td>
-                                <td>
-                                    <kbd>⌘</kbd>+<kbd>U</kbd>
-                                </td>
-                            </tr>
-                            <tr>
-                                <td>End of Page</td>
-                                <td>
-                                    <kbd>CTRL</kbd>+<kbd>I</kbd>
-                                </td>
-                                <td>
-                                    <kbd>⌘</kbd>+<kbd>I</kbd>
-                                </td>
-                            </tr>
-                        </tbody>
-                    </Table>
-                </Modal.Body>
-                <Modal.Footer className="modal-background" >
-                    <Button className="btn-sm" variant="secondary" onClick={handleCloseHelp}>
-                        Close
-                    </Button>
-                </Modal.Footer>
+                contentClassName={(THEME_STATES.LIGHT === theme)?"modal-light":"modal-dark"}>
+                <HelpModal handleClose={handleCloseHelp}></HelpModal>
             </Modal>
         </>
     );
